feat(oas31): render JSON Schema 2020-12 in Operations and Webhooks
Refs #8513
