+++ src/core/plugins/json-schema-2020-12/components/JSONSchema/JSONSchema.jsx
@@ -11,6 +11,7 @@ import {
   useLevel,
   useFn,
   useIsEmbedded,
+  useIsExpanded,
   useIsExpandedDeeply,
   useIsCircular,
   useRenderedSchemas,
@@ -24,8 +25,9 @@ import {
 const JSONSchema = forwardRef(
   ({ schema, name, dependentRequired, onExpand }, ref) => {
     const fn = useFn()
+    const isExpanded = useIsExpanded()
     const isExpandedDeeply = useIsExpandedDeeply()
-    const [expanded, setExpanded] = useState(isExpandedDeeply)
+    const [expanded, setExpanded] = useState(isExpanded || isExpandedDeeply)
     const [expandedDeeply, setExpandedDeeply] = useState(isExpandedDeeply)
     const [level, nextLevel] = useLevel()
     const isEmbedded = useIsEmbedded()
@@ -77,6 +79,8 @@ const JSONSchema = forwardRef(
     const KeywordReadOnly = useComponent("KeywordReadOnly")
     const KeywordWriteOnly = useComponent("KeywordWriteOnly")
     const ExpandDeepButton = useComponent("ExpandDeepButton")
+
+    console.dir(expandedDeeply)
 
     /**
      * Effects handlers.
