+++ src/core/plugins/json-schema-2020-12/components/keywords/Title/Title.jsx
@@ -9,6 +9,9 @@ import { useFn } from "../../../hooks"
 
 const Title = ({ title, schema }) => {
   const fn = useFn()
+  const renderedTitle = title || fn.getTitle(schema)
+
+  if (!renderedTitle) return null
 
   return (
     <div className="json-schema-2020-12__title">
