+++ src/core/plugins/oas31/components/_all.scss
@@ -1,0 +1,2 @@
+@import './model/model';
+@import './models/models';
