+++ src/core/plugins/oas31/components/models/_models.scss
@@ -1,0 +1,3 @@
+.models .json-schema-2020-12:not(.json-schema-2020-12--embedded) > .json-schema-2020-12-head .json-schema-2020-12__title:first-of-type {
+  font-size: 16px;
+}
