+++ src/style/main.scss
@@ -19,4 +19,5 @@
     @import 'split-pane-mode';
     @import 'markdown';
     @import '../core/plugins/json-schema-2020-12/components/all';
+    @import '../core/plugins/oas31/components/all';
 }
